organize custom metrics
